Extended chunk ID to 16-bit, added chunk limit property.

+++ SonLVLAPI/LayoutFormat.cs
@@ -61,8 +61,8 @@ namespace SonicRetro.SonLVL.API
 			else
 			{
 				LevelData.Log("Layout file \"" + filename + "\" not found.");
-				layout.FGLayout = new byte[DefaultSize.Width, DefaultSize.Height];
-				layout.BGLayout = new byte[DefaultSize.Width, DefaultSize.Height];
+				layout.FGLayout = new ushort[DefaultSize.Width, DefaultSize.Height];
+				layout.BGLayout = new ushort[DefaultSize.Width, DefaultSize.Height];
 				if (HasLoopFlag)
 				{
 					layout.FGLoop = new bool[DefaultSize.Width, DefaultSize.Height];
@@ -121,7 +121,7 @@ namespace SonicRetro.SonLVL.API
 			else
 			{
 				LevelData.Log("FG layout file \"" + filename + "\" not found.");
-				layout.FGLayout = new byte[DefaultSize.Width, DefaultSize.Height];
+				layout.FGLayout = new ushort[DefaultSize.Width, DefaultSize.Height];
 				if (HasLoopFlag)
 					layout.FGLoop = new bool[DefaultSize.Width, DefaultSize.Height];
 			}
@@ -134,7 +134,7 @@ namespace SonicRetro.SonLVL.API
 			else
 			{
 				LevelData.Log("BG layout file \"" + filename + "\" not found.");
-				layout.BGLayout = new byte[DefaultSize.Width, DefaultSize.Height];
+				layout.BGLayout = new ushort[DefaultSize.Width, DefaultSize.Height];
 				if (HasLoopFlag)
 					layout.BGLoop = new bool[DefaultSize.Width, DefaultSize.Height];
 			}
